# docs/api/openapi.yaml
openapi: "3.0.3"
info:
  title: CodeForge API
  description: |
    REST API for CodeForge -- AI coding agent orchestration platform.

    All endpoints under /api/v1 require a valid JWT bearer token unless
    explicitly marked as public (login, setup-status, health).

    Refresh tokens are delivered as httpOnly cookies on the /auth path.
  version: "0.8.0"
  license:
    name: AGPL-3.0-or-later
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
  - url: http://localhost:8080
    description: Local development (root, for /health)
...
paths:
...
  /llm/models:
    get:
      tags: [LLM]
      summary: List active LLM models
      description: Returns all models currently configured in LiteLLM.
      operationId: listLLMModels
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Model list
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    model_name:
                      type: string
                      example: "openai/gpt-4o"
                    litellm_params:
                      type: object
                      properties:
                        model:
                          type: string
                        api_base:
                          type: string
        "500":
          $ref: "#/components/responses/InternalServerError"
    post:
      tags: [LLM]
      summary: Add an LLM model to LiteLLM
      description: Platform admin only (an admin of the default tenant; models are shared by all tenants), audited. Others get 403.
      operationId: addLLMModel
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [model_name, litellm_params]
              properties:
                model_name:
                  type: string
                  example: "openai/gpt-4o"
                litellm_params:
                  type: object
                  properties:
                    model:
                      type: string
                    api_key:
                      type: string
                    api_base:
                      type: string
      responses:
        "201":
          description: Model added
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "ok"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /llm/models/{id}:
    delete:
      tags: [LLM]
      summary: Delete an LLM model deployment from LiteLLM
      description: |
        Platform admin only (an admin of the default tenant), audited. `id` is the LiteLLM
        deployment ID (`model_id` in the model list, taken from LiteLLM's `model_info.id`).
        Models are shared by all tenants.
        Replaces `POST /llm/models/delete` (405 since 2026-09-30).
      operationId: deleteLLMModel
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Model deleted
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "502":
          description: LiteLLM rejected the deletion
...
  /policies:
    get:
      tags: [Policies]
      summary: List policy profiles
      description: Returns all policy profiles (built-in presets and custom).
      operationId: listPolicyProfiles
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Policy profile list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PolicyProfile"
    post:
      tags: [Policies]
      summary: Create a custom policy profile
      description: Requires admin or editor role.
      operationId: createPolicyProfile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, rules]
              properties:
                name:
                  type: string
                description:
                  type: string
                rules:
                  type: array
                  items:
                    type: object
                    properties:
                      pattern:
                        type: string
                        description: Glob pattern matching tool calls
                      effect:
                        type: string
                        enum: [allow, deny, ask]
      responses:
        "201":
          description: Policy profile created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PolicyProfile"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: The name is a built-in preset or a policy file for it already exists
  /policies/allow-always:
    post:
      tags: [Policies]
      summary: Persist an "Allow Always" rule for a project
      description: |
        Admin only. Adds an allow rule for the tool (Bash: `command_allow` with every executable of the
        approved command) to the per-project clone `{profile}-custom-{project_id}` of the profile that
        decided the call. The clone replaces that profile only for this project; the project's profile
        selection is not changed. The rule never overrides a deny list. Persisted to `policy.custom_dir`.
      operationId: allowAlwaysPolicy
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [project_id, tool]
              properties:
                project_id:
                  type: string
                profile:
                  type: string
                  description: Profile named by the agui.permission_request event (optional; default is the project's effective profile)
                tool:
                  type: string
                  description: Tool name as sent by the worker (canonicalized server-side)
                command:
                  type: string
                  description: The approved bash command (Bash only)
      responses:
        "200":
          description: The updated profile clone
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PolicyProfile"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Unknown project or profile
        "409":
          description: No policy directory configured (policy.custom_dir is empty)
  /policies/{name}/evaluate:
    post:
      tags: [Policies]
      summary: Evaluate a tool call against a policy profile (policy tester)
      description: |
        Evaluates the call like a run would, in a synthetic workspace `/workspace` (the tester has no
        project): relative paths and Bash redirection targets are placed under it, an absolute path only when
        it is under `/workspace`.
      operationId: evaluatePolicy
      security:
        - bearerAuth: []
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tool]
              properties:
                tool:
                  type: string
                command:
                  type: string
                path:
                  type: string
      responses:
        "200":
          description: The decision and its reason
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"

  # -----------------------------------------------
  # Benchmarks (dev-mode only)
  # -----------------------------------------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access token obtained from POST /auth/login.
        Pass as: Authorization: Bearer <token>
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    Forbidden:
      description: Insufficient permissions (role check failed)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    PolicyProfile:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        rules:
          type: array
          items:
            type: object
            properties:
              pattern:
                type: string
              effect:
                type: string
                enum: [allow, deny, ask]
        built_in:
          type: boolean
...